More portability fixes

+++ tests/integration/teos-grify/test.sh
@@ -11,7 +11,7 @@ testMovieGrify(){
   dc-tools::assert::equal "$exit" "0"
   #dc-tools::assert::equal "$result" ""
 
-  rm "tests/integration/teos-grify/movie.mp4"
+  [ ! -f "tests/integration/teos-grify/movie.mp4" ] || rm "tests/integration/teos-grify/movie.mp4"
 
   [ "$(command -v ffmpeg)" ] || endSkipping
 }
